Add JSON serialization
The SegmentMap.getJson() function is implemented, that returns the JSON
representation of the map. The JSON is represented as follows:

{"<id-1>":
  {"id": <id-1>,
   "data": "<data-1>",
   "links": [<id-2>],
  },
 "<id-2>":
  {"id": <id-2>,
   "links": [],
  }
}

A custom serializer is written for Segment, because other segments
referenced in links should only be represented by their id (saves
space).

The links are also changed to a simplex representation to form a DAG.
Also saves space and avoids infinite loops.