Add tests and fix implementation
Hierarchical tests are added for SegmentCollection.

SegmentCollection.toJson() is subject to change once it is standardized,
now the implementations return different answers.

Instead of applying the adapter pattern in SegmentList, now an actual
backing list is contained in the class. This is done because ArrayList
does some uncontrollable things to backing list (extending, etc.).